requirements.txt
autopep8==2.0.4
certifi==2023.7.22
charset-normalizer==3.2.0
dnspython==2.4.2
idna==3.4
pycodestyle==2.11.0
requests==2.31.0
urllib3==2.0.4
